Much more error reporting, resolving TODOs

+++ src/sdf.zig
@@ -299,7 +299,7 @@ pub const SystemDescription = struct {
         pub const Vcpu = struct {
             id: u8,
             /// Physical core the vCPU will run on
-            cpu: ?u16 = null,
+            cpu: ?u8 = null,
         };
 
         pub fn create(allocator: Allocator, name: []const u8, vcpus: []const Vcpu, options: Options) !VirtualMachine {
